+++ src/dataPointSettings.ts
@@ -1,0 +1,6 @@
+"use strict";
+
+export class DataPointSettings {
+    // Table configuration
+    public tableConfiguration: string = "";
+}