+++ src/settings.ts
@@ -1,14 +1,9 @@
 "use strict";
 
 import { dataViewObjectsParser } from "powerbi-visuals-utils-dataviewutils";
+import { DataPointSettings } from "./dataPointSettings";;
 import DataViewObjectsParser = dataViewObjectsParser.DataViewObjectsParser;
 
 export class VisualSettings extends DataViewObjectsParser {
-      public dataPoint: dataPointSettings = new dataPointSettings();
-      }
-
-    export class dataPointSettings {
-      // Table configuration
-      public tableConfiguration: string = "";
-     }
-
+      public dataPoint: DataPointSettings = new DataPointSettings();
+}